requirements.txt

streamlit==1.27.2
streamlit-drawable-canvas==0.9.3
pillow
numpy
